backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
motor==3.5.1
pymongo==4.8.0
pydantic==2.9.2
python-dotenv==1.0.1
firebase-admin==6.5.0
python-jose[cryptography]==3.3.0
